+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This is the changelog for Cldr_units v3.16.0 released on ______, 2022.  For olde
 
 ### Enhancements
 
+* Add a `Cldr.Unit.Range` type to allow creating, enumerating and formatting unit ranges.
+
 * Add support for multiplication and division of arbitrary units. Thanks to @maennchen for the report (and patience). Closes #32.
 
 * Adjust the inspect protocol implementation to emit executable code strings.